==> 00b_extract_LMF_2b_track_info.ipynb <==
import pandas as pd
import os
from spotipy.oauth2 import SpotifyClientCredentials
import json
import spotipy
import time
from os import walk
import signal


from datetime import datetime

from tqdm.notebook import tqdm

SPOTIPY_CLIENT_ID=['e1be06382c2d486fb229713b05850f64',
                   '67ac6bda2f78458f8553018c3082340b',
                   'e1a52feaafdd43ad9eff6dccd21a966e',
                   'a4fdae2cec0c4755bd305bc64a0aa7ea',
                   '91dbf78e2bc34a2f87464f5fdd27faa9',
                   '5e9ea2ed76c14333919dda3094056355',
                   '38980db96ebb4f70b8875e51ec43936a',
                   '0e3d3bc703714da4ae3c7056435fa7f4',
                  '470500d793ae4e10ac394e655b542432',
                    '933647a0ae0344c1a01d3f7f4d512cb0']
SPOTIPY_CLIENT_SECRET=['412a96bba9444861874dbe27bfe13084',
                       'e9da73b0c4a849b8a62f726c3d015811',
                       '7ec98fe3c4be48dd818fcc3704c461bf',
                       '82c0243a290340c8b1b06f0de595622e',
                       '64c6ad405e8744ed8a3009a58e36e672',
                       '01e7bba1e37d4c6b8710f81a3ca49725',
                       '4acc1efc8e4a4265bda6be4594329c83',
                       '2aebaed63e55430c915f6be24ace0c9e',
                       'c110cbdd33ab49fca353046162c9ecfe',
                      'f482e0a7c8314b9d95d8719996853aef']

n_spotify_clients=len(SPOTIPY_CLIENT_ID)

client_index=5
client_credentials_manager = SpotifyClientCredentials(client_id=SPOTIPY_CLIENT_ID[client_index], 
                                                      client_secret=SPOTIPY_CLIENT_SECRET[client_index])
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager, requests_timeout= 5)

tracks_df= pd.read_csv(os.path.join('data', 'LMF-2b', 'spotify-uris.tsv'), sep='\t')
tracks_df

filenames = []
for (dirpath, dirnames, f) in walk(os.path.join('data', 'generated', 'LMF_2b')):
    filenames.extend(f)
    break
filenames

df_lst=[]
for f in filenames:
    df= pd.read_csv(os.path.join('data', 'generated', 'LMF_2b', f), index_col=0)
    df_lst.append(df)
processed_tracks_df= pd.concat(df_lst, axis=0)    

processed_tracks_df

processed_tracks_id_lst= processed_tracks_df['track_id'].unique()
total_track_id_lst= tracks_df['uri'].unique()

tracks_id_to_process_lst= list(set(total_track_id_lst) - set(processed_tracks_id_lst))
tracks_id_to_process_lst 

len(tracks_id_to_process_lst)

track_batch_size= 50
n_batches= len(tracks_id_to_process_lst) // track_batch_size

# Register an handler for the timeout
def time_out_handler(signum, frame):
    print("Spotify call timeout!")
    raise Exception("end of time")
    
signal.signal(signal.SIGALRM, time_out_handler)

track_info_lst=[]

counter=0
for i in tqdm(range(0, n_batches)):
    counter += 1
    track_ids= tracks_id_to_process_lst[i*track_batch_size:(i+1)*track_batch_size]
    
    signal.alarm(10) # Fire exeption after 10 seconds
    query_successful= True
    try:  
        tracks_info= sp.tracks(track_ids)
    except:
        signal.alarm(0)
        query_successful= False
        if len(track_info_lst)>0:
            print("Saving intermediate results...", end='')       
            now = datetime.now()
            dt_string = now.strftime("%Y%m%d_%H%M%S")
            aux_track_info_df= pd.DataFrame(track_info_lst, columns='track_id artists release_date release_date_precision'.split())
            aux_track_info_df.to_csv(os.path.join('data', 'generated', 'LMF_2b', f'00b_LFM_2B_tracks_{dt_string}.csv'))
            print('DONE')
        else:
            print("No intermediate results to save")
            
        track_info_lst=[]       
        counter=0
        
        #Refresh client to use for the extraction
        client_index = (client_index + 1) % n_spotify_clients
        print(f"Using new client  #{client_index}/{n_spotify_clients} ({SPOTIPY_CLIENT_ID[client_index]}, {SPOTIPY_CLIENT_SECRET[client_index]})")
        client_credentials_manager = SpotifyClientCredentials(client_id=SPOTIPY_CLIENT_ID[client_index], 
                                                              client_secret=SPOTIPY_CLIENT_SECRET[client_index])
        
        sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager, requests_timeout= 5)
    
    #Cancel alarm if spotify call return successfully
    if query_successful:
        signal.alarm(0)
        for track_info in tracks_info['tracks']:
            try:
                track_artists_lst=[]
                for artist in track_info['artists']:
                    track_artists_lst.append((artist['id'], artist['name']))
                release_date= track_info['album']['release_date']
                release_date_precision= track_info['album']['release_date_precision']
                track_info_lst.append((track_info['id'], track_artists_lst, release_date, release_date_precision))
            except:
                print("Error retrieving song data")

now = datetime.now()
dt_string = now.strftime("%Y%m%d_%H%M%S")
aux_track_info_df= pd.DataFrame(track_info_lst, columns='track_id artists release_date release_date_precision'.split())
aux_track_info_df.to_csv(os.path.join('data', 'generated', 'LMF_2b', f'00b_LFM_2B_tracks_{dt_string}.csv'))
 

print("That's all folks!")
