==> 04_merge_spotify_lfm_data.ipynb <==
import pandas as pd
import numpy as np
import os
from ast import literal_eval
from tqdm.notebook import tqdm

## Merge tracks data

audio_features_df= pd.read_csv(os.path.join('data', 'songs_features_2017_2022_v8.csv'), index_col=0)
audio_features_df=audio_features_df['id danceability energy key loudness mode speechiness acousticness instrumentalness liveness valence tempo'.split()]
audio_features_df

filenames = []
for (dirpath, dirnames, f) in os.walk(os.path.join('data', 'generated', 'spotify')):
    filenames.extend(f)
    break

df_lst=[]
for f in filenames:
    df= pd.read_csv(os.path.join('data', 'generated', 'spotify', f), index_col=0, converters={"artists": literal_eval})
    df_lst.append(df)
spotify_df= pd.concat(df_lst, axis=0) 
spotify_df= spotify_df.reset_index(drop=True)
spotify_df= spotify_df.drop_duplicates('track_id')
spotify_df

filenames = []
for (dirpath, dirnames, f) in os.walk(os.path.join('data', 'generated', 'LMF_2b')):
    filenames.extend(f)
    break

df_lst=[]
for f in filenames:
    df= pd.read_csv(os.path.join('data', 'generated', 'LMF_2b', f), index_col=0, converters={"artists": literal_eval})
    df_lst.append(df)
lfm_df= pd.concat(df_lst, axis=0) 
lfm_df= lfm_df.reset_index(drop=True)
lfm_df= lfm_df.drop_duplicates('track_id')
lfm_df

tracks_df= pd.concat([lfm_df, spotify_df], axis=0)
tracks_df= tracks_df.drop_duplicates('track_id')
tracks_df= tracks_df.reset_index(drop=True)

def get_artist_component(artist_list, component_index):
    artist_id_lst=[]
    for a in artist_list:
        artist_id_lst.append(a[component_index])
    return artist_id_lst

tracks_df['artist_id']= tracks_df['artists'].apply(lambda x: get_artist_component(x, 0))
tracks_df['artist_name']= tracks_df['artists'].apply(lambda x: get_artist_component(x, 1))
tracks_df= tracks_df.drop(columns='artists')

tracks_df['release_date_precision'].value_counts()

tracks_df= tracks_df[tracks_df['release_date_precision']!='year']
tracks_df= tracks_df[~tracks_df['release_date_precision'].isna()]

def normalize_release_date_format(row):
    release_date= row['release_date']
    if release_date is not None:
        if row['release_date_precision']=='month':
            release_date = release_date + '-15'
    return release_date

tracks_df['release_date']=tracks_df.apply(normalize_release_date_format, axis=1)

tracks_df

tracks_df.info()

tracks_w_features= tracks_df.merge(audio_features_df, how='left', left_on='track_id', right_on='id')
tracks_w_features= tracks_w_features.drop(columns='id')
tracks_w_features

tracks_w_features['danceability'].isna().sum()

tracks_w_features.to_csv(os.path.join('data','generated', '04_lfm_spotify_tracks_w_features.csv'))

## Merge artist data

def read_artist_info(source):
    filenames = []
    for (dirpath, dirnames, f) in os.walk(os.path.join('data', 'generated', source, 'artist_info_V2')):
        filenames.extend(f)
        break
    filenames
    
    df_lst=[]
    if len(filenames)>0:
        for f in filenames:
            df= pd.read_csv(os.path.join('data', 'generated', source, 'artist_info_V2', f), index_col=0)
            df_lst.append(df)
    artist_info_df= pd.concat(df_lst, axis=0)
    
    return artist_info_df

lfm2b_artist_info_df= read_artist_info('LMF_2b')
lfm2b_artist_info_df.head()

spotify_artist_info_df= read_artist_info('spotify')
spotify_artist_info_df.head()

all_artist_info= pd.concat([lfm2b_artist_info_df,spotify_artist_info_df], axis=0)
all_artist_info

type(all_artist_info.loc['2eJkpFtB62y7qCx27DDNGY', 'genre_5'])

artist_lst=[]
for artist_code, artist_info in tqdm(all_artist_info.iterrows()):
    genres=[]
    for c in 'genre_4 genre_5 genre_6 genre_7 genre_8 genre_9 genre_10 genre_11 genre_12 genre_13'.split():
        value= artist_info[c]
        if type(value)==str:
            genres.append(artist_info[c])
    artist_lst.append((artist_code, artist_info['artist_name'], artist_info['popularity'], artist_info['followers'], genres))

artist_info_compacted= pd.DataFrame(artist_lst, columns='artist_code artist_name popularity followers genres'.split())
artist_info_compacted= artist_info_compacted.drop_duplicates('artist_code')

artist_info_compacted

artist_info_compacted.to_csv(os.path.join('data','generated', '04_lfm_spotify_artist_info_V2.csv'))

## Add genres to the tracks

artist_info_= artist_info_compacted.set_index('artist_code')

artist_info_

def extract_genres(artist_code_lst):
    genres_lst=[]
    for ac in artist_code_lst:
        if ac in artist_info_.index:
            genres_lst= genres_lst + artist_info_.loc[ac,'genres']

    return genres_lst

def generate_genres_str(genres_lst):
    genres_lst_str=';'
    if len(genres_lst)>0:
        genres_lst_str= ','.join(genres_lst)
        genres_lst_str= ';' + genres_lst_str
    return genres_lst_str

tracks_w_features['genres']= tracks_w_features['artist_id'].apply(lambda x: extract_genres(x))
tracks_w_features['num_genres']= tracks_w_features['genres'].apply(lambda x: len(x))
tracks_w_features['genres_str']= tracks_w_features['genres'].apply(lambda x: generate_genres_str(x))

tracks_w_features

tracks_w_features[tracks_w_features['num_genres']>0]

tracks_w_features.to_csv(os.path.join('data','generated', '04_lfm_spotify_tracks_w_features_and_genres.csv'))

print("That's all folks!")
