==> 05b_merge_audio_features.ipynb <==
import pandas as pd
import os
from ast import literal_eval

all_tracks= pd.read_csv(os.path.join('data', 'generated', '04_lfm_spotify_tracks_w_features.csv'),
                        converters={"artist_id": literal_eval,
                                    "artist_name": literal_eval}, index_col=0)
all_tracks.shape

tracks_no_feat_df= all_tracks[all_tracks['danceability'].isna()]
tracks_no_feat_df= tracks_no_feat_df.set_index('track_id')
tracks_no_feat_df= tracks_no_feat_df['release_date release_date_precision artist_id artist_name'.split()]
tracks_no_feat_df

filenames = []
for (dirpath, dirnames, f) in os.walk(os.path.join('data', 'generated', 'audio_features')):
    filenames.extend(f)
    break
    
df_lst=[]
for f in filenames:
    df= pd.read_csv(os.path.join('data', 'generated', 'audio_features', f), index_col=0)
    df_lst.append(df)
audio_feat_df= pd.concat(df_lst, axis=0)
audio_feat_df= audio_feat_df.rename(columns={'id':'track_id'})
audio_feat_df= audio_feat_df.set_index('track_id')
audio_feat_df

audio_feat_df.shape

tracks_no_feat_df['danceability energy key loudness mode speechiness acousticness instrumentalness liveness valence tempo'.split()]= audio_feat_df['danceability energy key loudness mode speechiness acousticness instrumentalness liveness valence tempo'.split()]
tracks_no_feat_df= tracks_no_feat_df.reset_index()

tracks_no_feat_df

tracks_w_feat_df= all_tracks[~all_tracks['danceability'].isna()]
tracks_w_feat_df

all_tracks_w_feat_df= pd.concat([tracks_w_feat_df,tracks_no_feat_df ], axis=0)
all_tracks_w_feat_df= all_tracks_w_feat_df.reset_index(drop=True)
all_tracks_w_feat_df= all_tracks_w_feat_df.dropna(axis=0)
all_tracks_w_feat_df

all_tracks_w_feat_df.isna().sum()

all_tracks_w_feat_df

all_tracks_w_feat_df['release_date'].min(), all_tracks_w_feat_df['release_date'].max()

all_tracks_w_feat_df.to_csv(os.path.join('data', 'generated', '05b_all_tracks_w_feat_df.csv'))

### Add column indicating the track's genres

artist_info_df= pd.read_csv(os.path.join('data','generated', '04_lfm_spotify_artist_info_V2.csv'), 
                         converters={"genres": literal_eval}, index_col=0)
artist_info_df

artist_info_= artist_info_df.set_index('artist_code')
artist_info_

def extract_genres(artist_code_lst):
    genres_lst=[]
    for ac in artist_code_lst:
        if ac in artist_info_.index:
            genres_lst= genres_lst + artist_info_.loc[ac,'genres']
    return genres_lst

def generate_genres_str(genres_lst):
    genres_lst_str=';'
    if len(genres_lst)>0:
        genres_lst_str= ','.join(genres_lst)
        genres_lst_str= ';' + genres_lst_str
    return genres_lst_str

all_tracks_w_feat_df['genres']= all_tracks_w_feat_df['artist_id'].apply(lambda x: extract_genres(x))
all_tracks_w_feat_df['num_genres']= all_tracks_w_feat_df['genres'].apply(lambda x: len(x))
all_tracks_w_feat_df['genres_str']= all_tracks_w_feat_df['genres'].apply(lambda x: generate_genres_str(x))

all_tracks_w_feat_df= all_tracks_w_feat_df.drop_duplicates('track_id')

all_tracks_w_feat_df

all_tracks_w_feat_df[all_tracks_w_feat_df['num_genres']>0]

all_tracks_w_feat_df.to_csv(os.path.join('data', 'generated', '05b_all_tracks_w_feat_and_genre.csv'))

print("That's all folks!")
